Move SQLite journal mode pragma to connection string
A small change to move the pragma to activate `journal_mode` to the
connection string instead of being run separately as its own `PRAGMA`
command. It should be a tiny bit faster, and it's not harmful for
legibility (in fact, I think it improves it in a small way too).