Test support for LibSQL using SQLite driver (#957)
LibSQL is a SQLite fork designed to be a little more to progress and
open-source contribution, and it seems to have picked up some real world
usage. They're trying to be careful in maintaining API compatibility
with SQLite, so we should be able to support it as long as they continue
to do so.

As of #955, we can now add testing for LibSQL without having to bring it
in as a project dependency. That gets pushed down to `riverdrivertest`
instead.

I put the new example test in `riverdrivertest` as well so that
`riversqlite` doesn't need to add a libSQL dependency. For consistency,
I also moved the SQLite example test to `riverdrivertest` as well so
they're both easy to find. A nice byproduct of that move is that
`riversqlite` doesn't need a dependency on `modernc.org/sqlite`, which
is kind of nice. It lets you bring your own SQLite driver.