Implement `JobDeleteMany` operations
Here, add functions for `Client.JobDeleteMany` which lets jobs be
deleted in batches. The main impetus for this is to give us a way of
implementing a "purge queue" function, which has been previously
requested by users, and which is generally just a good feature.

We copy the `JobList` API very closely and use almost all the same
implementation, with the caveat that I've removed some of the more
complex features that I don't think will be as necessary for deletion,
with the sorting features being the biggest one, but also arbitrary SQL
in `WHERE` queries. These can always be added in later because they're
purely additive to the API, but I'm hoping that we won't need them.