Have the reindexer reindex more `river_job` indexes
Here, expand the scope of the reindexer so that beyond the pair of
indexes it reindexes right now, it also reindexes all the other ones in
the `river_job` table. This won't make a huge different for many uses of
River, but it'll make a particular difference in cases where for some
amount of time `river_job` became very large, but has since contracted
to a much more modest size. B-trees expand but never shrink, so in that
case the indexes will be left very large despite holding very little,
and the only way to get them back down to size is to rebuild them.

More reindexes will always put more load on the database, but we're
still doing that as carefully as we can in that indexes are rebuilt one
by one, and using `CONCURRENTLY`. Back in #935 we added protection so
that in case of a situation where index builds are too expensive and
start to fail, the reindexer will bail out instead of trying and failing
over and over again.