Accept `river migrate-down --target-version 0` to remove all River tables (#966)
There was a bit of an unfortunate UX quirk in the CLI wherein to remove
all River tables you needed to specify a special value of
`--target-version -1`. This is a relic of how the CLI uses `rivermigrate`
internally, and `rivermigrate`'s options have a `TargetVersion int` that
makes 0 a reserved value for when the option is unset.

It seems like we can do a little better here for the CLI, so here we add
some code that lets 0 or -1 be used as a `--target-version` to apply all
down migrations. We do this by changing the CLI default value to the
sentinel value -2, which we change back to a real default before
forwarding onto `rivermigrate`.