Have `HookWorkEnd.WorkEnd` receive a `JobRow` in addition to an error
Here, modify the `HookWorkEnd.WorkEnd` so that it receives a `JobRow` of
the job that just finished working in addition to an error. It's not
great having to change this, but having a `JobRow` to work with will be
crucial for most functionality that people might like to use this hook
to implement.

It's a breaking change, but hopefully not a bad one because I think not
having `JobRow` previously is a sign that not many people had yet tried
to implement anything with `HookWorkEnd` yet, because they would've
needed a `JobRow` for most of what they wanted to do.