Let logger be overridden from `rivertest.Worker`
Currently, `rivertest.Worker` always used a log based on the input
`testing.T` test case, ignoring a logger that's passed in via
configuration. This isn't normally a problem, but an be in cases like if
a user is trying to discard all logging like described in #979.

Here, let the logger be overridden by passing in one explicitly via the
River configuration struct.

Fixes #979.