Allow retention periods to be configured to -1 so they're retained indefinitely (#990)
Here, let each of the job cleaner's configured retention periods
(`CancelledJobRetentionPeriod`, `CompletedJobRetentionPeriod`,
`DiscardedJobRetentionPeriod`) accept the special value of -1 so that
the cleaner will maintain that type of job indefinitely, effectively
disabling it.

This probably won't be of much use day-to-day for most users, but does
enable us to augment cleaning functionality with special overrides.