Remove `WorkerDefaults.Hooks`
Remove `WorkerDefaults.Hooks`. I believe this was a vestige left in
after a number of refactoring passes as it's not used anywhere (if I
recall correctly, workers may have started with hooks, but they were
moved to `JobArgs` because hooks the job args part of an args/worker
combination are more available in more places).

This is technically a breaking change, but this was never used at any
point and it shouldn't realistically break anyone.