Fewer driver allocations in client tests + resolve TODOs (#1007)
Here, a very minor change to resolve a few TODOs around not being able
to reuse allocated drivers due to use of test helpers. The fix here is
that since `newTestClient` is just a thin wrapper around `NewClient`
anyway, just drop its use in favor of `NewClient` in a couple places.