Schedule jobs in smaller batches when it appears the database is degraded
The job scheduler currently tries to work in batches that are quite
large, trying to schedule 10,000 jobs at once. Normally this is fine
because there's far fewer than 10,000 jobs to schedule, or if there are
many jobs you're generally on a large Postgres instance where scheduling
a lot of jobs at once should be fine.

However, in a situation where Postgres is severely degraded, it might be
possible that get yourself in a situation where the scheduler can never
succeed. It tries to schedule 10k jobs, fails due to timeout, tries
again, fails again, etc. Jobs accumulate and the situation keeps getting
worse. You can see an example of this in #995, where the users in
question ran out of IOPS on their DB and saw their situation steadily
degrade.

Here, put in a partial stab at an approach for ameliorating this. We add
a circuit breaker to the scheduler such that if its scheduling operation
fails three times in a row due to timeout, the circuit breaks, and the
scheduler switches from its default batch of 10k to a much reduced batch
of 1,000 jobs. The timeout must occur three times in a row for this to
occur. If timeouts are intermittent (i.e. in the case of an only
partially degraded database), the breaker is allowed to reset, and
operations continue to use the 10k limit.

An executive decision I made here is to have the circuit breaker stay
open once it trips, so after the scheduler switches to a reduced batch
of 1k, it'll continue using the 1k batch until the program restarts. The
reason I designed it this way is that in my experience after a database
becomes degraded, it tends to stay degraded over the long term. We could
have it so that it switches back to 10k job batches after some time, but
more likely than not those 10k operations will just start failing again,
and make things worse by putting more load on the database while they
do. I could see tweaking or reexamining this behavior in the future, but
I don't think either way we could go here is that much more correct than
the other.

I've only applied this to the job scheduler for this initial design, but
I'd like to eventually bring this to all/most of the other maintenance
services, many of which haven't had a huge amount of thought put into
how their batch size works since they were first introduced.