Add lazy slice formatters to `slogutil` (#1028)
Adds a couple slice helpers to `slogutil` that implement `slog.LogValue`
to achieve the effect of formatting slices that are used as values in
slog attributes, but only when we actually need to do so because a slog
log line is being printed. Otherwise, `LogValue` is never called and we
avoid the associated work and allocations.