Require use of `JobDeleteManyParams.UnsafeAll` to delete all jobs
Addresses #1031. It's come to light that it's relatively easy to
accidentally delete all non-running jobs by invoking `JobDeleteMany`
with `nil` parameters or parameters with all defaults.

Here, make it a little harder to do so by requiring an invocation to
`JobDeleteManyParams.UnsafeAll` before a blanket delete is allowed.
Calling `UnsafeAll` isn't necessary if any filters are applied.

Fixes #1031.