Cleanly handle jobs being snoozed with zero duration
As I was going through to see if we could recommend snoozing jobs with a
zero duration as a way of returning an error that'd cause them to be
retried immediately but without logging error noise, I found that
snoozes with zero duration do work, but had a bug in the subscription
manager that'd cause them to be emitted as `EventKindJobFailed` instead
of `EventKindJobSnoozed`

Here, go through and fix that problem, and making sure to add adequate
test coverage for it.

I also add a formal documentation to `river.JobSnooze` that suggests the
function can be used for a noise-less retry even in cases where no
snooze time is required.