fix flaky job snooze 0 duration test
The test was fetching a job immediately after it finished execution with
a snooze duration of 0. However, sometimes it could be picked up for
work again prior to being loaded for inspection.

To avoid this, look only at the job row coming in from the subscription
since we cannot otherwise know how soon the client will fetch again.