Test against Postgres 18 (#1049)
Very small one to add Postgres 18, which was released last week, to the
test matrix.

I didn't drop an old version out because I'm not sure what our policy on
that is yet, but we should probably drop old databases only with more
discretion. I'm not sure we want to go all the way with Postgres formal
policy of support [1] (under which Postgres 13 is still supported until
November), but we should probably be fairly generous since databases are
still hard to upgrade, and maybe do a little user outreach before
dropping old versions.

[1] https://www.postgresql.org/support/versioning/