Add test case checking for snoozed event kind on zero duration snooze
Quick follow up to add regression coverage for #1057. We already had a
test case in there that checks for snooze events being emitted on a more
normal snooze with non-zero snooze duration, but the problem being fixed
only occurs on zero duration snoozes. Here, add one more test case that
covers that as well.

I also looked into extracting the event-kind-from-job-completer-result
logic so that it could be shared between client and test worker and
prevent this sort of problem down the line, but it's unfortunately made
quite difficult because `river.EventKind*` lives in `river` instead of
`rivertype`, which means that it can't be referenced without making the
top-level `river` a dependency. And if `river` becomes a dependency, a
cyclic dependency is introduced.

Also added a changelog for #1057.