Add `RollbackWithoutCancel` helper to have rollbacks supersede context cancellation (#1062)
Here, add a `dbutil.RollbackWithoutCancel` helper. The purpose of this
is in situations where an operation is in a transaction but ends up
cancelling due to a context timeout. From there, there tends to be a
rollback in a `defer`, but the rollback ends up not running because its
context is already cancelled.

The new helper removes a cancelled context, then adds a new 5 second
context timeout to run the rollback command. This is mainly used for
`JobSchedule` in the hopes of coming up with a fix for #1059, but we use
the new function everywhere that we previously had a manual rollback,
and it also gets add to the `dbutil.WithTx*` helpers so that it gets
used anywhere those do (and these should probably be preferred most of
the time because it makes rollbacks and commits harder to forget).

Fixes #1059.