Allow removing periodic jobs by ID (#1071)
Here, we add two new functions to `PeriodicJobBundle`:

* `RemoveByID`
* `RemoveManyByID`

These are in addition to the existing `Remove` and `RemoveMany`.

The new variants are useful in cases where durable jobs are in use.
Tracking jobs by ID rather than handle may be more convenient, and
provides a stable identifier across multiple runs. IDs were added after
jobs could be added or removed dynamically, so even where durable jobs
aren't in use, it may still be more convenient to remove them by ID.

Unlike `Remove`, I've made `RemoveByID` return a boolean indicating
whether the removal happened or not. We may want to add this to `Remove`
as well given it's a useful feature, but I didn't do so for now as
adding a return value could be considered a minor breaking change.

Notably in the case of durable periodic jobs, removing a job by ID
doesn't immediately remove it from the database. This still relies on
the existing system of allowing unconfigured durable periodic jobs
records to be pruned over time. We'll see if a database removal has to
be a hard requirement from those requesting this feature, but not
needing a database call makes things considerably simpler because
(1) there's no risk of contention if a job is removed across multiple
clients, and (2) no error return is necessary.