Correct uniqueness annotations for fields in substructs
This one's aimed at fixing #1075. A deficiency in the previous unique
field implementation is that only top-level args fields were considered.
Anything unique annotations in a field substruct or an embedded
substruct were skipped.

Here, we change the reflect algorithm so that it descends into structs
recursively, similar to the behavior of `encoding/json` or other
reflect-based packages. See the new test cases for the specifics of how
this works.

We try to maintain backwards compatibility so that if a substruct is
tagged with `river:"unique"`, but none of the fields in the substruct
are, we use the entire JSON-encoded value of the substruct for
uniqueness calculations, exactly like how it worked before this change.
If previously there were `river:"unique"` annotations on subfields,
behavior will change, but I think this is for the best because as
described in #1075, River was almost certainly not doing what the user
would've expected, thereby possibly creating an accidental bug.