Add hook for when the periodic job enqueuer starts up
This one's an alternative (and hopefully a better one) to #1078. #1078
works for removing a periodic job from a leader, but doesn't cover the
case where a new periodic job might need to be added dynamically from
any possible client.

Here, we add the new hook `HookPeriodicJobsStart` which runs when the
periodic job enqueuer starts up on a newly elected leader. This gives
the hook quite a bit of power to dynamically adjust all currently
configured periodic jobs on the enqueuer. For example, it may clear all
of them and then add them back from scratch using a canonical set of
periodic job records from elsewhere which it knows should exist.

This will need to be paired with another feature that would allow users
to force a leader reelection (which I haven't written yet). So when they
determine that a periodic job overhaul is necessary, they'll knock the
current leader offline and then use `HookPeriodicJobsStart` to configure
the right set of periodic jobs.

A modification we make here is to make it possible to get a client from
context in the hook similar to how it's possible in a worker right now.
This is because the hook implementation will often want access to a
client instance, but it's very awkward getting access to a client
otherwise because you define your hooks before a client is available,
thereby creating a chicken and egg problem.