Unique args: Handle embedded field that is *not* a struct
Aims to fix #1087.

One thing I hadn't accounted for is that it's actually possible to have
an embedded type that's *not* a struct. e.g.

    type MyStruct struct {
        string // embedded non-struct
    }

The simplified example above is *not* because we short circuit unless
properties are exported (and `string` is not exported because it starts
with a lower case letter), but if we have a similar case with a custom
type (e.g. `type MyString string`) then we can reproduce the error
described in #1087.

Fix the problem by making sure to check that a field is really a struct
before recursing back into `getSortedUniqueFields`. We're able to
simplify the function somewhat by handling all structs (anonymous or
not) in one spot since the logic is so similar.

Fixes #1087.