Add `riverlog.LoggerSafely` for access outside work body without panic (#1093)
Here, add a "safe" variant of `riverlog.Logger`: `rigerlog.LoggerSafely`.

While it'd be rare to not know whether you had a logger or not from
inside a `Work` function, `Work` functions might call into other Go
modules in your project, those modules might be reused in non-`Work`
contexts, and it might not always be clear whether a logger should be
available or not.

This function lets users do a conditional check before logging:

    if logger, ok := riverlog.LoggerSafely(ctx); ok {
        logger.InfoContext(ctx, "Hello from logger")
    }

I went with the "safely" naming according to our conventions elsewhere,
but one thing I'm not sure about is that `ClientFromContextSafely` which
is a similar function, returns an error instead of `ok` boolean. Here we
break with that convention somewhat, but I think the boolean version is
more ergonomic. Another option might be to give this a slightly
alternative naming convention like `LoggerOK` or `LoggerMaybe`.

Fixes #1092.