Move request resign to "bundle style" functions
Move away from generic `Notify`/`NotifyTx` functions and over to a
"bundle style" convention similar to what we have on `PeriodicJobs`:

    client.Notify().RequestResign(ctx)
    client.Notify().RequestResignTx(ctx, tx)

The nice part about this pattern is that we get the benefits of both
alternative styles. We no longer need the generic kind any `any` payload
that we previously had on `Notify(...)`, but we also get to organize all
notifications in one place so they're not all top level client functions
that would make the API overly crowded.