`Notify().*` functions for client + force resign mechanism (#1085)
* `Notify`/`NotifyTx` functions for client + force resign mechanism

This one's the counterpart to #1084: a mechanism that requests that the
current River leader resign. Users can use this to force a resignation
and then `HookPeriodicJobsStart` from #1084 to dynamically reconfigure
all periodic jobs the next time a leader is elected.

Introduces two new functions to the client:

* `Notify`
* `NotifyTx`

The idea is that this will be used to force resign for now, but could be
expanded to other types of manual notifications that need to be sent.

The alternative would be add something like `LeaderResign`, but it'd be
nice to have some naming that implies that we're _requesting_ via
Postgres listen/notify that the leader resign rather than that we're
specifically requesting that _this_ client resigns (because this client
may not be the leader). So maybe something like `NotifyLeaderResign`
could be plausible, but that's not particularly great either. The other
thing to remember is that whatever we choose, we need to have a *tx
variant like `NotifyLeaderResignTx`.

* Move request resign to "bundle style" functions

Move away from generic `Notify`/`NotifyTx` functions and over to a
"bundle style" convention similar to what we have on `PeriodicJobs`:

    client.Notify().RequestResign(ctx)
    client.Notify().RequestResignTx(ctx, tx)

The nice part about this pattern is that we get the benefits of both
alternative styles. We no longer need the generic kind any `any` payload
that we previously had on `Notify(...)`, but we also get to organize all
notifications in one place so they're not all top level client functions
that would make the API overly crowded.