Add test showing round trip out job output (#1102)
I was working on another job output related change, found myself
becoming quite confused by the way it was implemented, and further found
that it doesn't have much in the way of testing through `Output` (only a
couple unit tests on `JobRow`) so it was hard to verify that what I was
changing was producing the expected result.

Here, add a "round trip" on job output for the client that records
output in a work function, then retrieves the row and checks `Output` on
it.