Introduce `Client.JobUpdate` function that can store output incrementally (#1098)
Here, take a stab at a solution for #1064. We introduce a new client
function `Client.JobUpdate` that takes any output currently recorded in
context and stores it to the given job row. `JobUpdate` currently only
sets output, but the idea is that it could be expanded in the future in
case it's useful to do so.

The reason that we don't have a `river.PersistOutput` in line with
`river.RecordOutput` is that once we're talking about storing data,
we're dealing with the usual persistence semantics. We need a client
instance with access to an executor, and we probably want to have a
`*Tx` variant like we have for all other functions like this one.

Fixes #1064.