Bump golangci-lint to 2.7.2 (#1107)
I upgrade golangci-lint recently and I guess it's been a while since we
did it in River because I got a whole bunch of new errors. Most are
pretty perfunctory, but to be fair, it did catch quite a few docstrings
where one of our symbols was named somewhat wrong at the beginning.