Initialize `nil` migration opts earlier (#1117)
The migration functions are supposed to allow you to call them with
nil options, but I noticed in some cases it's not working properly for
`MigrateTx` and it's possible to raise a nil pointer panic.

Here, initialize a nil `opts` earlier to prevent the problem. This makes
the code more obvious/conventional anyway in that it's easier to see
that nil options is supposed to be supported.